fix:MetricsCollectTask has cycle dependence
